Change: Separate eir module into many sources
