+++ src/ir/identifier.rs
@@ -1,0 +1,1 @@
+pub type Identifier = String;
