+++ src/ir/mod.rs
@@ -1,6 +1,14 @@
+pub mod binding_kind;
 pub mod display;
-pub mod eir;
+pub mod identifier;
+pub mod literal;
+pub mod node;
 pub mod type_;
+pub mod value;
 
-pub use self::eir::{BindingKind, Identifier, Literal, Node, Value};
+pub use self::binding_kind::BindingKind;
+pub use self::identifier::Identifier;
+pub use self::literal::Literal;
+pub use self::node::Node;
 pub use self::type_::Type;
+pub use self::value::Value;
